+++ CHANGELOG.md
@@ -29,6 +29,7 @@ like `Source Id`, `Message Id` or `Instance` will have this set to true.
 - #486: Missing identity designation in the canboat schema
 - #500: PGN 129540 parameter Range residuals has incorrect resolution and unit
 - #506: 0x1ED00 and 0x1EE00 are not single frame but fast packet
+- #504: Document handling of max and two bit numbers
 
 ### Added